Made code a bit more advanced
Created some new functions, added function pointers, and saved some
memory space by changing a lot of ints to shorts

+++ src/board.h
@@ -53,7 +53,7 @@ extern const char *eraser[];
  * 'win_height' specifies the height for each WINDOW.
  * 'win_width' specifies the width for each WINDOW.
  */
-void win_layout(board_square *win_grid, int rows, int cols, int win_height, int win_width);
+void win_layout(board_square *win_grid, short rows, short cols, short win_height, short win_width);
 
 /**
  * Create a "wall" by reversing the foreground and background colors of
@@ -70,7 +70,7 @@ void win_layout(board_square *win_grid, int rows, int cols, int win_height, int 
  *      2 - Wall is on upper side of WINDOW.
  *      3 - Wall is on lower side of WINDOW.
  */
-void draw_wall(WINDOW *win, int placement);
+void draw_wall(WINDOW *win, short placement);
 
 /**
 * Take in an array of strings, print those strings to a WINDOW, and center them 
@@ -88,7 +88,7 @@ void draw_wall(WINDOW *win, int placement);
 * 'pair' specifies which color pair to format the image in.
 * 'win_pair' is the color pair for the WINDOW (used to find the background color).
 */
-void win_draw_image(WINDOW *win, const char **image, int img_height, short pair, short win_pair);
+void win_draw_image(WINDOW *win, const char **image, short img_height, short pair, short win_pair);
 
 /**
  * Place a new WINDOW on a specified side of a WINDOW currently on the board. The color
@@ -107,6 +107,6 @@ void win_draw_image(WINDOW *win, const char **image, int img_height, short pair,
  *      2 - Above the neighboring WINDOW.
  *      3 - Below the neighboring WINDOW.
  */
-WINDOW *place_win(WINDOW *neighbor, int placement, int win_height, int win_width);
+WINDOW *place_win(WINDOW *neighbor, short placement, short win_height, short win_width);
 
 #endif    // _BOARD_H
